Expand macros testing for register scalars, arrays
The test cases proposed in the register_array tests should be covering
every kind of usage pattern. But, it was strictly for arrays of
registers. Since I'm about to change the macros, I wanted to expand that
for scalars, and for the various types of instances that we could
supply.

This commit implements those common test cases for scalar registers and
arrays of registers. The common test cases carry over the shadowed
variable tests. The common test cases are expressed in yet another
macro, making it easy to test register scalars and arrays with the same
inputs and asserts. The trouble using macros to generate cases is that
you lose error localization with the assert_eq! macros. There's assert
strings to help mitigate this issue.