De-duplicate macros for scalar, array support
An indirection layer extracts a scalar register, then supplies it to
another macro. That second macro describes the read and / or write on
that scalar register.

I used the test cases from the previous commit to test. I also
integrated and tested within my RAL.

The previous commit revealed another way to support register arrays
without this indirection. See the child commit for details. I'm keeping
this as an intermediate commit just in case we prefer this approach.