Support register arrays with zero or more offsets
Figured this out when implementing the common test cases. We can extend
the existing macro patterns with an example that matches zero or more
bracket-enclosed numbers. This commit reverts the parent commit, and
extends the original patterns with the new example.

A new pathological test case shows that we can support multi-dimensional
arrays.